raise errors for execute statements
The error handling for executeCommand existed, but executeCommand had a
flag to consume errors and always return success.
With these adjustments it is now possible to adjust the settings in a
way that execute errors are handled as errors and therefore can
influence the processing of clusters.

To make it backwards compatible, the default is as before. It must be
enabled to get an effect.